Move implementation detection inside components
