+++ packages/apollos-ui-connected/src/utils/nodeImplementsType/nodeImplementsType.tests.js
@@ -1,0 +1,17 @@
+// eslint-disable-next-line no-unused-vars
+import ApollosConfig from '@apollosproject/config';
+import nodeImplementsType from '.';
+
+jest.mock('@apollosproject/config', () => ({
+  TYPEMAP: {
+    WeekendContentItem: ['Node', 'Feature'],
+  },
+}));
+
+describe('nodeImplementsType', () => {
+  test('correctly detects implementation', () => {
+    const nodeId = 'WeekendContentItem:abcd1234';
+    const result = nodeImplementsType(nodeId, 'Node');
+    expect(result).toBeTruthy();
+  });
+});
